Overview: Pushing revisions 27-47 from https://testoriented.googlecode.com/svn/branches/smart-templates
Enhances the test generation mechanism to produce richer test method stubs that contain local variables to represent parameters, a call to the method under test and a call to assertEquals with the return value of said call, if appropriate.

Changes:
Consult the individual change logs in the "smart-templates" branch for details.


Issue: 3